// src/app/interfaces/servs.ts
export interface Servs   {
    id: number;
   name_serv: string;
   fecha:Date;
   descripcion: string;
   recomendaciones: string;
   partes: string;
   horas_serv:number;
   equipo_id: number;
   
}

export class Servicio  {
    id: number;
   name_serv: string;
   fecha:Date;
   descripcion: string;
   recomendaciones: string;
   partes: string;
   horas_serv:number;
   equipo_id: number;
   
}